Add files via upload
This revision of the top case modifies the tab geometry so that the tabs fit more securely into the recesses in the bottom case. I moved the face of each tab outward by 0.25mm and also modified the angle of the tabs by 10 degrees. This prevents the top case from accidentally popping off the bottom case when removing the A/V adapter from the C64. Changes were made by modifying the existing .stl in Solidworks.